requirements.txt
flask==2.3.3
google-cloud-translate==2.0.1
google-cloud-language==2.11.1